requirements.txt
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.0.3
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
numpy==2.1.1
packaging==24.1
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
tzdata==2024.2
Werkzeug==3.0.4
